Move classes into separate files, post list view
